+++ examples/crypt/terraform.tf
@@ -2,11 +2,7 @@ terraform {
   required_providers {
     flexbot = {
       source  = "igor-feoktistov/flexbot"
-      version = ">=1.6.1"
-    }
-    rancher2 = {
-      source = "rancher/rancher2"
-      version = ">= 1.10.6"
+      version = ">=1.6.2"
     }
   }
   required_version = ">= 0.13"