+++ examples/crypt/variables.tf
@@ -1,0 +1,3 @@
+variable "pass_phrase" {
+  type = string
+}