+++ examples/rancher-server-flexbot/terraform.tf
@@ -2,11 +2,11 @@ terraform {
   required_providers {
     flexbot = {
       source  = "igor-feoktistov/flexbot"
-      version = ">=1.6.1"
+      version = ">=1.6.2"
     }
     rke = {
       source = "rancher/rke"
-      version = ">= 1.1.5"
+      version = ">= 1.1.7"
     }
   }
   required_version = ">= 0.13"