+++ examples/rancher-workload-cluster-flexbot/terraform.tf
@@ -4,6 +4,10 @@ terraform {
       source  = "igor-feoktistov/flexbot"
       version = ">=1.6.2"
     }
+    rancher2 = {
+      source = "rancher/rancher2"
+      version = ">= 1.11.0"
+    }
   }
   required_version = ">= 0.13"
 }