+++ examples/repo/versions.tf
@@ -2,7 +2,7 @@ terraform {
   required_providers {
     flexbot = {
       source  = "igor-feoktistov/flexbot"
-      version = ">=1.6.1"
+      version = ">=1.6.2"
     }
   }
   required_version = ">= 0.13"